Update README.md
Package Scanner is a Python tool that allows you to scan a directory, look through it, and compile a list of all packages used in Python files. It can generate a `requirements.txt` file with the list of packages that can be installed from pip, as well as other packages found.